Reorganize 17 billion things!
The frontend files are now all in src/frontend
..except for the 'assets' files which are still where they were
..except for the css/ which was under assets but logically should be in frontend/.
The server is now in backend/, and there's a small README there.
This also moves geojson.py into scripts/ and documents it, and commits cleanups I should have committed last month.

TODO: this committ removes run.py. We need to recreate it.