Guiding light comments.
The Autobahn API is in flux and I don't have time to fix this example properly.
In fact, I don't know if I *should* fix it; I did intentionally include a copy
of autobahn.js to isolate this example from the flux.
But of course now that flux has made this example apocryphal as written currently...