Finally make the graphs play nice with the map.
Also, disable Crafty by default, because crafty interferes with standard css.

It turned out that the sample code Majd copied for the d3 graph was
very very
very picky. It had a hardcoded width in it and its container was
float: right
which is why it was over top: it was floating and too large not to.
The current solution is not ideal--it uses jquery to set the width instead of CSS
and it still uses floats
but at least it shows what I want, and it can be used as a baseline.
(see http://stackoverflow.com/questions/446060/css-two-divs-next-to-each-other for gore)