Silly me, who needs *<form>s* to use HTML widgets anymore?
I am so daft. I obviously grew up in another era, when forms were forms
and buttons always submitted or reset.

Also make the slider even smooooooother, since there's no harm (at this point) in doing so.
Eventually this will be pegged back to integers corresponding to timesteps, and the output
labelling--but only the labelling--will map timesteps to dates.