[FIXED] Repair eutopia to the same outputs as before
I switched to using what I'm calling my "version 1" database schema:
runID | time | duramWheat | duramWheatGreen | ....
which simplifies a lot of things, but seems unSQLish.

Though I think simulation.py is still broken.
and server.py is definitely going to need a fix.

Also chmod +x ./eutopia.py, for convenience and to point
out to the unixheads that *it is itself a program*