SQLAlchemy(as in, not Dataset)-based magic model logger.
SQLAlchemy has a hook for providing default values, which we exploit to the fullest;
we win the ability for the user to explicitly set a run_id or time if they really want to,
and need not even override insert() to do it.

This version doesn't actually have a ModelLog class. That will have to be thought about.